requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
pandas>=2.0.0
openpyxl>=3.1.0
matplotlib>=3.8.0
streamlit>=1.40.0
python-dotenv>=1.0.0
